Move to FriendsOfFlarum

+++ migrations/2020_07_31_06_create_discussion_answer_table.php
@@ -5,16 +5,20 @@ use Illuminate\Database\Schema\Builder;
 
 return [
     'up' => function (Builder $schema) {
-        $schema->create('flagrow_mason_discussion_answer', function (Blueprint $table) {
+        if ($schema->hasTable('fof_mason_discussion_answer')) {
+            return;
+        }
+
+        $schema->create('fof_mason_discussion_answer', function (Blueprint $table) {
             $table->unsignedInteger('discussion_id');
             $table->unsignedInteger('answer_id');
             $table->timestamps();
 
             $table->foreign('discussion_id')->references('id')->on('discussions')->onDelete('cascade');
-            $table->foreign('answer_id')->references('id')->on('flagrow_mason_answers')->onDelete('cascade');
+            $table->foreign('answer_id')->references('id')->on('fof_mason_answers')->onDelete('cascade');
         });
     },
     'down' => function (Builder $schema) {
-        $schema->drop('flagrow_mason_discussion_answer');
+        $schema->drop('fof_mason_discussion_answer');
     },
 ];
